chore

+++ src/main/java/com/quickticket/quickticket/domain/creditTransaction/entity/CreditTransaction.java
@@ -25,7 +25,7 @@ public class CreditTransaction {
 
     @NotNull
     @Column(nullable = false)
-    private Integer changeAmount;
+    private Long changeAmount;
 
     @NotNull
     @Column(nullable = false)
